added . for libft

+++ core/str/ft_lstrlen.c
@@ -6,18 +6,18 @@
 /*   By: viwade <viwade@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/27 02:48:54 by viwade            #+#    #+#             */
-/*   Updated: 2019/05/21 07:31:51 by viwade           ###   ########.fr       */
+/*   Updated: 2019/05/27 12:20:32 by viwade           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft.h"
 #include <string.h>
 
-size_t	ft_lstrlen(const unsigned *s)
+int64_t	ft_lstrlen(const int *s)
 {
-	unsigned	*e;
+	int	*e;
 
-	e = (unsigned*)s;
+	e = (int *)s;
 	while (*(e += !!e[0]))
 		;
 	return ((e - s) / sizeof(int));
