+++ ext/added_functions/ft_arrlen.c
@@ -6,13 +6,13 @@
 /*   By: viwade <viwade@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/13 19:19:31 by viwade            #+#    #+#             */
-/*   Updated: 2019/05/13 19:08:15 by viwade           ###   ########.fr       */
+/*   Updated: 2019/05/27 12:21:04 by viwade           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft.h"
 
-size_t	ft_arrlen(char **ar)
+int64_t	ft_arrlen(char **ar)
 {
 	unsigned long long int	len;
